Extract platform console printing helper
