+++ frb_rust/src/misc/frb_logging.rs
@@ -51,9 +51,7 @@ macro_rules! enable_frb_rust_to_dart_logging {
                 }
 
                 let Ok(sink) = self.sink.read() else {
-                    $crate::for_generated::frb_logging_fallback_print(
-                        "FRB logging sink lock is poisoned",
-                    );
+                    $crate::for_generated::print_to_console("FRB logging sink lock is poisoned");
                     return;
                 };
                 let Some(sink) = sink.as_ref() else {
@@ -68,7 +66,7 @@ macro_rules! enable_frb_rust_to_dart_logging {
                     file: record.file().map(ToOwned::to_owned),
                     line: record.line(),
                 }) {
-                    $crate::for_generated::frb_logging_fallback_print(&format!(
+                    $crate::for_generated::print_to_console(&format!(
                         "Failed to forward Rust log to Dart: {error:?}",
                     ));
                 }
@@ -142,75 +140,4 @@ macro_rules! enable_frb_rust_to_dart_logging {
             $setup_dart_logging_output
         }
     };
-}
-
-#[doc(hidden)]
-pub fn frb_logging_fallback_print(message: &str) {
-    #[cfg(target_os = "android")]
-    {
-        frb_logging_fallback_print_android(message);
-        return;
-    }
-
-    #[cfg(any(target_os = "ios", target_os = "macos"))]
-    {
-        frb_logging_fallback_print_apple(message);
-        return;
-    }
-
-    #[cfg(target_family = "wasm")]
-    {
-        crate::for_generated::web_utils::js_console_error(message);
-        return;
-    }
-
-    #[cfg(not(any(
-        target_os = "android",
-        target_os = "ios",
-        target_os = "macos",
-        target_family = "wasm"
-    )))]
-    {
-        eprintln!("{message}");
-    }
-}
-
-#[cfg(target_os = "android")]
-fn frb_logging_fallback_print_android(message: &str) {
-    use std::ffi::CString;
-    use std::os::raw::{c_char, c_int};
-
-    const ANDROID_LOG_ERROR: c_int = 6;
-
-    #[link(name = "log")]
-    extern "C" {
-        fn __android_log_write(prio: c_int, tag: *const c_char, text: *const c_char) -> c_int;
-    }
-
-    let tag = CString::new("flutter_rust_bridge")
-        .expect("static Android log tag should not contain interior null");
-    let text = frb_logging_to_c_string(message);
-
-    unsafe {
-        __android_log_write(ANDROID_LOG_ERROR, tag.as_ptr(), text.as_ptr());
-    }
-}
-
-#[cfg(all(any(target_os = "ios", target_os = "macos"), feature = "user-utils"))]
-fn frb_logging_fallback_print_apple(message: &str) {
-    oslog::OsLog::new("flutter_rust_bridge", "logging").error(message);
-}
-
-#[cfg(all(
-    any(target_os = "ios", target_os = "macos"),
-    not(feature = "user-utils")
-))]
-fn frb_logging_fallback_print_apple(message: &str) {
-    eprintln!("{message}");
-}
-
-#[cfg(target_os = "android")]
-fn frb_logging_to_c_string(message: &str) -> std::ffi::CString {
-    std::ffi::CString::new(message.replace('\0', "(null)"))
-        .expect("interior nulls should be replaced")
 }
