+++ frb_rust/src/misc/mod.rs
@@ -5,6 +5,7 @@ pub(crate) mod into_into_dart;
 pub(crate) mod logs;
 pub(crate) mod manual_impl;
 pub(crate) mod panic_backtrace;
+pub(crate) mod printing;
 #[cfg(feature = "user-utils")]
 pub(crate) mod user_utils;
 pub(crate) mod version;
